+++ lib/core/utils/constants.dart
@@ -1,21 +1,11 @@
 class Assets {
-  static const String ADOBEXD = 'assets/img/adobe-xd.png';
+  static const String ADOBEXD = 'assets/img/adobe_xd.png';
   static const String ANDROID = 'assets/img/android.png';
   static const String FLUTTER = 'assets/img/flutter.png';
   static const String GIT = 'assets/img/git.png';
-  static const String SPOTIFY = 'assets/img/spotify.png';
-  static const String GOOGLE_PODCASTS = 'assets/img/google_podcasts.png';
-  static const String ANCHOR = 'assets/img/anchor.png';
-  static const String APPLE_PODCAST = 'assets/img/apple_podcasts.png';
-  static const String DEEZER = 'assets/img/deezer.png';
-  static const String POCKET_CASTS = 'assets/img/pocket_casts.png';
-  static const String CASTBOX = 'assets/img/castbox.png';
-  static const String RADIOPUBLIC = 'assets/img/radiopublic.png';
-  static const String OVERCAST = 'assets/img/overcast.png';
-  static const String BREAKER = 'assets/img/breaker.png';
 }
 
 class Urls {
-  static const String PODCAST = 'https://anchor.fm/s/127cd640/podcast/rss';
+  static const String PODCAST = 'https://universoflutter.com';
   static const String PROFILE = 'https://deandreamatias.com/profile.png';
 }